+++ cmd/tools/tools.go
@@ -26,6 +26,7 @@ import (
 	"github.com/streamingfast/firehose-core/cmd/tools/fix"
 	"github.com/streamingfast/firehose-core/cmd/tools/mergeblock"
 	print2 "github.com/streamingfast/firehose-core/cmd/tools/print"
+	"github.com/streamingfast/firehose-core/cmd/tools/relayer"
 	"github.com/streamingfast/logging"
 	"go.uber.org/zap"
 )
@@ -79,6 +80,7 @@ func ConfigureToolsCmd[B firecore.Block](
 	ToolsCmd.AddCommand(mergeblock.NewToolsUnmergeBlocksCmd(chain, logger))
 	ToolsCmd.AddCommand(mergeblock.NewToolsMergeBlocksCmd(chain, logger))
 	ToolsCmd.AddCommand(fix.NewToolsFixBloatedMergedBlocks(chain, logger))
+	ToolsCmd.AddCommand(relayer.NewToolsRelayerGroup(chain, logger))
 
 	if chain.Tools.MergedBlockUpgrader != nil {
 		ToolsCmd.AddCommand(mergeblock.NewToolsUpgradeMergedBlocksCmd(chain, logger))